cvinsight_api/pyproject.toml
[project]
name = "cvinsight-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "openai>=2.36.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.13.4",
    "pydantic-settings>=2.14.0",
    "pymupdf>=1.27.2.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.27",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]
